sudokuTS-3 Implement the X-Wing Rule: Refactored.
Moved the abstract rules to their own class. Abstracted and refactored
the X-Wing rule.
